feat(auth): add audit mode and metrics for namespace_scope enforcement
Adds a per-provider rollout switch and observability so operators can stage
namespace_scope before hard denial:
- namespace_scope_enforcement = "enforce" (default, 403 on mismatch) or
  "audit" (allow the write but log and count it). Lets operators deploy,
  watch, then switch on enforcement.
- nora_auth_namespace_scope_total{provider,decision} counts every decision
  (allow | deny | would_deny), so a staged rollout is observable.

Also relocates NamespaceAuthority + enforce_namespace_scope from validation
to a new auth::namespace module: validation is compiled into the fuzz lib and
must stay free of the binary-only metrics/config deps. The pure segment-aware
namespace_match stays in validation. Handlers now import the enforcement entry
points from crate::auth; behavior is unchanged for the enforce default.

Part of #583
