+++ nora-registry/src/registry/cargo_registry.rs
@@ -12,12 +12,13 @@
 
 use crate::activity_log::{ActionType, ActivityEntry};
 use crate::audit::AuditEntry;
+use crate::auth::{enforce_namespace_scope, NamespaceAuthority};
 use crate::registry::{
     circuit_open_response, method_not_allowed, nora_base_url, proxy_fetch, ProxyError,
 };
 use crate::registry_type::RegistryType;
 use crate::secrets::expose_opt;
-use crate::validation::{enforce_namespace_scope, validate_storage_key, NamespaceAuthority};
+use crate::validation::validate_storage_key;
 use crate::AppState;
 use axum::{
     body::Bytes,
