+++ nora-registry/src/registry/pypi.rs
@@ -11,13 +11,14 @@
 
 use crate::activity_log::{ActionType, ActivityEntry};
 use crate::audit::AuditEntry;
+use crate::auth::{enforce_namespace_scope, NamespaceAuthority};
 use crate::registry::{
     circuit_open_response, method_not_allowed, nora_base_url, proxy_fetch, proxy_fetch_text,
 };
 use crate::registry_type::RegistryType;
 use crate::secrets::expose_opt;
 use crate::ui::components::html_escape;
-use crate::validation::{ends_with_ci, enforce_namespace_scope, NamespaceAuthority};
+use crate::validation::ends_with_ci;
 use crate::AppState;
 use axum::{
     extract::{Multipart, Path, State},
